Cleanup
- move generics to Generic namespace
- fix coding style
- workaround for FluentUI title focus bug
