+++ Debounce/Void.cs
@@ -1,0 +1,10 @@
+﻿// SPDX-FileCopyrightText: 2024 Frans van Dorsselaer
+//
+// SPDX-License-Identifier: MIT
+
+namespace Dorssel.Utilities;
+
+/// <summary>
+/// Void data type for <see cref="IDebouncer"/>.
+/// </summary>
+public readonly record struct Void { }
