+++ Examples/BlazorServerPush/Components/Pages/Counter.razor
@@ -12,8 +12,9 @@
 
 @using System.ComponentModel
 @using Dorssel.Utilities
+@using Dorssel.Utilities.Generic
 
-<h1>Counter</h1>
+<h2>Counter</h2>
 
 <p>Current count: @GlobalCounter.Count</p>
