+++ Examples/BlazorServerPush/Components/Pages/Home.razor
@@ -6,7 +6,7 @@
 
 @page "/"
 
-<h1>Demo of using Debouncer in blazor.</h1>
+<h2>Demo of using Debouncer in blazor.</h2>
 
 <p>This is a minor adaptation of the default blazor Counter demo that comes with the .NET SDK.</p>
