+++ Examples/Testable/TestableClass.cs
@@ -4,6 +4,7 @@
 
 using System.Diagnostics;
 using Dorssel.Utilities;
+using Dorssel.Utilities.Generic;
 using Void = Dorssel.Utilities.Void;
 
 namespace Testable;
