+++ UnitTests/Usings.cs
@@ -3,4 +3,5 @@
 // SPDX-License-Identifier: MIT
 
 global using Dorssel.Utilities;
+global using Dorssel.Utilities.Generic;
 global using Microsoft.VisualStudio.TestTools.UnitTesting;
